devices/device_simulator/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
aiohttp==3.8.5
asyncio-mqtt==0.13.0
aiocoap==0.4.7
coapthon3